requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
tqdm>=4.65.0
requests>=2.28.0
httpx[socks]>=0.28.0

